updated for various bug fixes around missing deployment slots, updated version, changed the publisher to polarissolutions per microsoft guidance

+++ Tasks/UpdateAzureConfigurationSetting/Update-AzureConfigurationSetting.ps1
@@ -57,25 +57,30 @@ function UpdateSettingInRoles([xml]$configuration, [string]$setting, [string]$va
 Write-Host "Updating setting $SettingKey for $ServiceName" -ForegroundColor Green
 
 # get current settings from Azure
-$deployment = Get-AzureDeployment -ServiceName $ServiceName -Slot $Slot -ErrorAction Stop
+$deployment = Get-AzureDeployment -ServiceName $ServiceName -Slot $Slot -ErrorAction SilentlyContinue
 
-$configuration = [xml]$deployment.Configuration
-
-# update setting if needed
-$updated = UpdateSettingInRoles $configuration $SettingKey $SettingValue
-
-if (-not($updated))
+if ($deployment)
 {
-    Write-Host "No settings have been updated as they are either up to date or not found"
-    return
+    $configuration = [xml]$deployment.Configuration
+
+    # update setting if needed
+    $updated = UpdateSettingInRoles $configuration $SettingKey $SettingValue
+
+    if (-not($updated))
+    {
+        Write-Host "No settings have been updated as they are either up to date or not found"
+        return
+    }
+
+    # save as a temporary file and upload settings to Azure
+    $filename = $env:temp + "\" + $ServiceName + ".cscfg"
+    $configuration.Save("$filename")
+
+    Set-AzureDeployment -Config -ServiceName $ServiceName -Configuration "$filename" -Slot $Slot
+
+    Remove-Item ("$filename")
+}else {
+    Write-Warning "No deployment was found.  Skipping configuration setting update..."
+    
 }
-
-# save as a temporary file and upload settings to Azure
-$filename = $env:temp + "\" + $ServiceName + ".cscfg"
-$configuration.Save("$filename")
-
-Set-AzureDeployment -Config -ServiceName $ServiceName -Configuration "$filename" -Slot $Slot
-
-Remove-Item ("$filename")
-
 Write-Host "Done"
